Pull latest from Honeycomb; 1.8.3 --> 2.1.0 (#4)
* Don't capture MongoDB session objects (#174)

* Don't capture MongoDB session objects

The MongoDB session object is passed into some calls as `options.session` and cannot be serialised to JSON, thus crashing the instrumentation.

Fixes #157

* Fix modification to parameters

I realised the previous fix would modify the parameter object `options` which might not be desired (or might be problematic). This version instead omits the `session` property from it when passing to `prefixKeys` instead of directly modifying the object.

* Remove old code

* PG instrumentation callbacks (#175)

Try to better handle the arguments being passed to the `query` function. It looks like having these three arguments have been pretty stable for many years. 

We need to decipher where in the arguments the callback lives. From reading the source, it looks like the options are these:

// query string, values, callback
// query string, callback
// Query like object
// Query like object, callback

Once we have deciphered where the correct callback lives we wrap it and pass it onto the wrapped `query` function.

This seems to work in my local manual testing. Looking into how to add some unit tests for this as well.

Resolves #173.

* 1.8.4

* Add lint step & fix lint error (#176)

Previously we only ran the lint step as part of the publish step which meant that any lint errors would block publishing.

This creates a new job that performs the linting before any of the test steps runs and removes the lint from the publish step.

In order to fix the existing lint error we had to set the ecma version to 2018.

* 1.8.5

* Everything must have filters :rage1:

* 1.8.6

* bump libhoney dep to ^2.0.2 (#178)

Bring in 2.0.2 of libhoney (https://github.com/honeycombio/libhoney-js/releases/tag/2.0.2) which drops (locally) `require()` time for the beeline from ~170ms to ~50ms.

Also take this time to drop documented support non-LTS node versions.

* 2.0.0

* [fastify] only wrap the request/reply hooks (#180)

Add a whitelist for request/reply hooks (https://github.com/fastify/fastify/blob/master/docs/Hooks.md), and just pass the others through unshimmed (the shim code assumes we're dealing with those hooks currently.)

Also make things a bit more robust by not crashing if there are requests that somehow didn't start a trace.

* 2.0.1

* [From geckoboard] fix errors when not in a trace (#181)

* Stop the beeline crashing if running outside of a trace

* Ensure we invoke async spans if the parent context isn't valid

Co-authored-by: Matt Button <matt.button@geckoboard.com>

* Remove entries from previously unbounded maps (#183)

* The request is safe to delete from finishersByRequest because
it is accessed only once, on the previous line, so long as there
is only one 'onResponse' per request.
* The request is safe to delete from trackedByRequest because
onResponse is the last available hook in the lifecycle.

* 2.0.2

* Bump acorn from 5.7.3 to 5.7.4 (#185)

* clean up a bunch of span/context related hacks and bugs (#184)

A few things changed here - one thing in a backward incompatible way _but_ in a way consistent with our docs.

1. Introduce a `Span` class, and change the api implementations to use instances of `Span` instead of regular JS objects in their `context.stack`.
2. `Span` exposes an `addContext` method that acts similar to the go beeline's -- no prefixing of `app.` when using it.  Change all instrumentation to call this method directly, since we always have a span there.
3. `Span` also has fields for `startTime`/`startTimeHR` so all the handling of those fields can be cleaned up - no more `delete`ing them from the payload when we send (and in tests!  ugh.)
3. Add a toplevel `addTraceContext` api method that acts similar to the go beeline's -- prefix with `app.` and also propagates to outbound http/https.
4. The backward incompatible change alluded to above: `api.addContext` no longer results in that context being propagated to outbound calls.  This is the documented behavior, so I don't feel as bad about making this change.

Now for the good stuff.  This PR also marks some things as deprecated or to-be-changed with the next major bump:

1. `api.customContext.{add, remove}` goes away entirely.  `.add` is the same as `api.addTraceContext`.  there is no replacement for `.remove`
2. similarly, `api.removeContext` goes away without replacement.
3. `api.addContext` will prefix with `app.`.  This mirrors the go beeline - if you have direct reference to a span you can call `span.addContext` and it won't prepend `app.`.  If you call the toplevel `addContext`, it will.

Added some more tests that exercise the various context adding methods so we'll fail those if we have regressions again in any of the documented behavior.

Fixes #149 
Fixes #156 
maybe others...

* parse AWS trace header (#187)

We care about Root/Parent clauses currently, and map those to traceId/parentSpanId.

This code matches the behavior of code aws uses in their sdks.

We'll need to figure out how propagation works in the aws world (on outbound requests) so we can add our custom stuff there as well, which will in turn cause more changes to this code.

algorithm for dealing with the header is basically:

1. split aws header by `;`
2. walk the list of strings from #1, splitting each string by `=` into [key, value].
    2a. if we see key="Root", value becomes traceId.
    2b. if we see key="Parent", value becomes parentSpanId.
3. If we didn't see traceId, bail.
4. If we didn't see parentSpanId, fall back to traceId.

* 2.1.0

* split('=') needs max=2 or we break base64 encoded context with padding (#191)

Fixes #190 

Splitting `k=v` strings by `=` fails if `v` contains an `=`, which is quite likely given that we're base64 encoding our trace `context`.  use `.split("=", 2)` to fix.

* Test that non-Honeycomb header trace parsing is configurable. (#194)

Make header constants consistent. Add tests for setting traceIdSource explicitly in the express of fastify initialization code. Users who do not want automatic parsing of incoming Amazon trace ID headers or Request IDs should now specify a traceIdSource in their initialization code to prevent missing root spans.

```
require("honeycomb-beeline")({
    writeKey:    "foo",
    dataset:     "bar",
    serviceName: "foo-service",
    express: {
        traceIdSource: 'X-Amzn-Trace-Id',
    },
})
```

In the next major release, we'll make this behavior the default (only Honeycomb headers will be parsed unless another header format is explicitly specified).

Co-authored-by: James Ross <twpol@users.noreply.github.com>
Co-authored-by: Martin Holman <martin@honeycomb.io>
Co-authored-by: Martin Holman <me@martinholman.co.nz>
Co-authored-by: Chris Toshok <toshok@hound.sh>
Co-authored-by: Chris Toshok <toshok@gmail.com>
Co-authored-by: Matt Button <matt.button@geckoboard.com>
Co-authored-by: Cameron Dunn <60160513+camerondunn-ns8@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Paul Osman <paul@honeycomb.io>